# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.ContactMessage:
    properties:
      attempts:
        type: integer
      createdAt:
        type: string
      email:
        maxLength: 255
        type: string
      id:
        type: integer
      lastError:
        type: string
      message:
        type: string
      name:
        maxLength: 255
        type: string
      sentAt:
        type: string
      status:
        type: string
      subject:
        maxLength: 500
        type: string
      updatedAt:
        type: string
    required:
    - email
    - message
    - name
    - subject
    type: object
  models.ContactMessageCreate:
    properties:
      email:
        maxLength: 255
        type: string
      message:
        maxLength: 10000
        type: string
      name:
        maxLength: 255
        type: string
      subject:
        maxLength: 500
        type: string
      website:
        description: Named "website" to trick bots
        type: string
    required:
    - email
    - message
    - name
    - subject
    type: object
  models.Recipient:
    properties:
      createdAt:
        type: string
      email:
        maxLength: 255
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      name:
        maxLength: 255
        type: string
      updatedAt:
        type: string
    required:
    - email
    - name
    type: object
  models.RecipientCreate:
    properties:
      email:
        maxLength: 255
        type: string
      isActive:
        description: Optional, defaults to true
        type: boolean
      name:
        maxLength: 255
        type: string
    required:
    - email
    - name
    type: object
  models.RecipientUpdate:
    properties:
      email:
        maxLength: 255
        type: string
      isActive:
        type: boolean
      name:
        maxLength: 255
        type: string
    type: object
host: localhost:8086
info:
  contact: {}
  description: API for contact form submissions and recipient management
  title: Messaging API
  version: "1.0"
paths:
  /api/v1/contact:
    post:
      consumes:
      - application/json
      description: Creates a new contact message (public endpoint)
      parameters:
      - description: Contact message
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.ContactMessageCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit a contact message
      tags:
      - Contact
  /api/v1/messages:
    get:
      description: Returns all contact messages (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ContactMessage'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all contact messages
      tags:
      - Messages
  /api/v1/messages/{id}:
    get:
      description: Returns a single contact message (admin only)
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ContactMessage'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get contact message by ID
      tags:
      - Messages
  /api/v1/recipients:
    get:
      description: Returns a list of all email recipients (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Recipient'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all recipients
      tags:
      - Recipients
    post:
      consumes:
      - application/json
      description: Creates a new email recipient (admin only)
      parameters:
      - description: Recipient data
        in: body
        name: recipient
        required: true
        schema:
          $ref: '#/definitions/models.RecipientCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Recipient'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new recipient
      tags:
      - Recipients
  /api/v1/recipients/{id}:
    delete:
      description: Deletes a recipient by ID (admin only)
      parameters:
      - description: Recipient ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a recipient
      tags:
      - Recipients
    get:
      description: Returns a single recipient by ID (admin only)
      parameters:
      - description: Recipient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipient'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get recipient by ID
      tags:
      - Recipients
    put:
      consumes:
      - application/json
      description: Updates an existing recipient (admin only)
      parameters:
      - description: Recipient ID
        in: path
        name: id
        required: true
        type: integer
      - description: Recipient data
        in: body
        name: recipient
        required: true
        schema:
          $ref: '#/definitions/models.RecipientUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipient'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a recipient
      tags:
      - Recipients
  /health:
    get:
      description: Returns service health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - Health
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"